@@ -8,7 +8,7 @@
     "AFL-3.0"
   ],
   "require": {
-    "php": "^7.1.0",
+    "php": ">=8.0",
     "flagshipcompany/flagship-api-sdk": "^1.1"
   },
   "autoload": {
